refactor: extract AwaImporter from JsbImportService
Create JsbParser/Importers/AwaImporter with import() and importFiles()
methods. JsbImportService now forwards processAwaData/processAwaFile
to the internal AwaImporter instance. RANK_SUFFIXES constant and
mb_convert_encoding CP1252→UTF-8 logic preserved verbatim.
